# lusevokteren-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FjordVind Lusevokteren API
  description: |
    REST API for sea lice monitoring in Norwegian aquaculture.

    ## Authentication
    Most endpoints require authentication via JWT Bearer token.

    ```
    Authorization: Bearer <your-token>
    ```

    ## Demo Mode
    In development, demo tokens are available:
    - `demo_token_admin` - Admin access
    - `demo_token_driftsleder` - Operations manager
    - `demo_token_røkter` - Fish farmer

  version: 1.1.0
  contact:
    name: FjordVind Support
    email: support@fjordvind.no
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.fjordvind.no
    description: Production server
...
paths:
...
  /api/treatments:
    get:
      summary: List treatments
      tags: [Treatments]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [planlagt, pågår, fullført, kansellert]
        - name: merd_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of treatments

    post:
      summary: Create treatment
      tags: [Treatments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTreatmentRequest'
      responses:
        '201':
          description: Treatment created
  /api/treatments/{id}:
    get:
      summary: Get treatment by ID
      tags: [Treatments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Treatment data
        '404':
          description: Treatment not found

    put:
      summary: Update treatment
      tags: [Treatments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTreatmentRequest'
      responses:
        '200':
          description: Treatment updated

  # ========== ALERTS ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateTreatmentRequest:
      type: object
      required:
        - treatment_type
        - scheduled_date
      properties:
        merd_id:
          type: string
          format: uuid
        locality_name:
          type: string
        treatment_type:
          type: string
          enum: [Hydrogenperoksid, Termisk, Mekanisk, Rensefisk, Ferskvann, Imidakloprid]
        scheduled_date:
          type: string
          format: date
        notes:
          type: string
    UpdateTreatmentRequest:
      type: object
      properties:
        status:
          type: string
          enum: [planlagt, pågår, fullført, kansellert]
        completed_date:
          type: string
          format: date
        lice_after:
          type: number
        effectiveness_percent:
          type: number
        notes:
          type: string
...